requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.4
aioredis==2.0.1
aiosignal==1.4.0
aiosmtplib==5.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
bcrypt==3.2.2
beanie==1.23.6
black==26.3.1
blinker==1.9.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
colorama==0.4.6
cryptography==46.0.6
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
fakeredis==2.34.1
fastapi==0.135.2
fastapi-mail==1.6.2
filetype==1.2.0
frozenlist==1.8.0
google-auth==2.49.1
google-genai==1.69.0
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
iniconfig==2.3.0
isort==8.0.1
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.13
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.23
langchain-google-genai==4.2.1
langchain-openai==1.1.12
langchain-text-splitters==1.1.1
langgraph==1.1.3
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langsmith==0.7.22
lazy-model==0.2.0
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.2
motor==3.3.1
multidict==6.7.1
mypy_extensions==1.1.0
numpy==2.4.4
openai==2.30.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
passlib==1.7.4
pathspec==1.0.4
platformdirs==4.9.4
pluggy==1.6.0
propcache==0.4.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
PyJWT==2.12.1
pymongo==4.6.1
pypdf==6.9.2
pytest==9.0.2
pytest-asyncio==1.3.0
python-docx==1.2.0
python-dotenv==1.2.2
python-multipart==0.0.22
pytokens==0.4.1
PyYAML==6.0.3
redis==7.4.0
regex==2025.11.3
requests==2.33.0
requests-toolbelt==1.0.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.48
starlette==0.52.1
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.42.0
watchfiles==1.1.1
websockets==16.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
